Alice/requirements.txt
discord.py==2.3.2
python-dotenv
nextcord
python-dotenv
colorama
ffmpeg-python
requests
loguru
PyNaCl
aiosqlite
Pillow